fix(lazycell): Fix unsound aliasing in `LazyCell::fill`
Fix unsound aliasing of the inner `Option<T>`. Per the discussion in
https://github.com/matklad/once_cell/pull/33, it is possible to obtain a
unique reference to the inner `Option<T>` despite it being potentially
aliased by a previous `.borrow()`. Kudos to @danielhenrymantilla for
spotting this issue in `matklad/once_cell` and to @matklad for pointing out that
it also applies to `LazyCell::fill`.

Closes #98.